+++ tests/test_settings.py
@@ -1553,7 +1553,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "1.9.1"
+    assert data.version == "1.9.2"
     assert (
         localization["quick_look_card_integration_group"]["en"]
         == "Mod Integration: Quick Look Card"